Create December-Sales.ipynb
# Data-Analytics
This is to demonstrate my passion for data analytics and data science . Working on projects everyday to get better at my skill and demonstrate here . 
This is a product sales dataset Currently done sorting on this. Split purchase address to have states,city,address and also display days of which purchase was made by using the date column to help me have more insights of what can be achieved with the dataset. 
Looking forward to answering with visualizations these questions:
1. The best selling items for this online shopping store.
2. Which states and cities purchase most.
3. Which Items sells most.
4. Which days does customers purchase most.
5.What times does customers purchase most and what they purchase.